Refactor SpecmaticConfig and update mapper name
