+++ core/src/main/kotlin/io/specmatic/core/config/SpecmaticConfig.kt
@@ -2,70 +2,20 @@ package io.specmatic.core.config
 
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.core.JsonParseException
-import com.fasterxml.jackson.databind.JsonMappingException
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
-import io.specmatic.core.Configuration.Companion.configFilePath
-import io.specmatic.core.Feature
-import io.specmatic.core.config.mapper.SpecmaticConfigMapper
 import io.specmatic.core.config.v1.SpecmaticConfigV1
 import io.specmatic.core.config.v2.ContractConfig
 import io.specmatic.core.config.v2.FileSystemConfig
 import io.specmatic.core.config.v2.GitConfig
 import io.specmatic.core.config.v2.SpecmaticConfigV2
-import io.specmatic.core.log.logger
-import io.specmatic.core.parseContractFileToFeature
-import io.specmatic.core.pattern.ContractException
 import io.specmatic.core.pattern.parsedJSONObject
 import io.specmatic.core.utilities.Flags
 import io.specmatic.core.utilities.Flags.Companion.EXAMPLE_DIRECTORIES
 import io.specmatic.core.utilities.Flags.Companion.getBooleanValue
 import io.specmatic.core.utilities.Flags.Companion.getStringValue
-import io.specmatic.core.utilities.exceptionCauseMessage
 import io.specmatic.core.value.Value
-import java.io.File
-import java.io.IOException
-
-
-const val APPLICATION_NAME = "Specmatic"
-const val APPLICATION_NAME_LOWER_CASE = "specmatic"
-const val CONFIG_FILE_NAME_WITHOUT_EXT = "specmatic"
-const val DEFAULT_TIMEOUT_IN_MILLISECONDS: Long = 6000L
-const val CONTRACT_EXTENSION = "spec"
-const val YAML = "yaml"
-const val WSDL = "wsdl"
-const val YML = "yml"
-const val JSON = "json"
-val CONFIG_EXTENSIONS = listOf(YAML, YML, JSON)
-val OPENAPI_FILE_EXTENSIONS = listOf(YAML, YML, JSON)
-val CONTRACT_EXTENSIONS = listOf(CONTRACT_EXTENSION, WSDL) + OPENAPI_FILE_EXTENSIONS
-const val DATA_DIR_SUFFIX = "_data"
-const val TEST_DIR_SUFFIX = "_tests"
-const val EXAMPLES_DIR_SUFFIX = "_examples"
-const val DICTIONARY_FILE_SUFFIX = "_dictionary.json"
-const val SPECMATIC_GITHUB_ISSUES = "https://github.com/znsio/specmatic/issues"
-const val DEFAULT_WORKING_DIRECTORY = ".$APPLICATION_NAME_LOWER_CASE"
-
-const val SPECMATIC_STUB_DICTIONARY = "SPECMATIC_STUB_DICTIONARY"
-
-const val MISSING_CONFIG_FILE_MESSAGE = "Config file does not exist. (Could not find file ./specmatic.json OR ./specmatic.yaml OR ./specmatic.yml)"
-
-fun invalidContractExtensionMessage(filename: String): String {
-    return "The file $filename does not seem like a contract file. Valid extensions for contract files are ${CONTRACT_EXTENSIONS.joinToString(", ")}"
-}
-
-fun String.isContractFile(): Boolean {
-    return File(this).extension in CONTRACT_EXTENSIONS
-}
-
-fun String.loadContract(): Feature {
-    if(!this.isContractFile())
-        throw ContractException(invalidContractExtensionMessage(this))
-
-    return parseContractFileToFeature(File(this))
-}
 
 data class SpecmaticConfig(
     val sources: List<Source> = emptyList(),
@@ -183,45 +133,5 @@ data class SpecmaticConfig(
             allPatternsMandatory = this.allPatternsMandatory,
             defaultPatternValues = this.defaultPatternValues
         )
-    }
-}
-
-fun loadSpecmaticConfigOrDefault(configFileName: String? = null): SpecmaticConfig {
-    return if(configFileName == null)
-        SpecmaticConfig()
-    else try {
-        loadSpecmaticConfig(configFileName)
-    }
-    catch (e: ContractException) {
-        logger.log(exceptionCauseMessage(e))
-        SpecmaticConfig()
-    }
-}
-
-fun loadSpecmaticConfig(configFileName: String? = null): SpecmaticConfig {
-    val configFile = File(configFileName ?: configFilePath)
-    if (!configFile.exists()) {
-        throw ContractException("Could not find the Specmatic configuration at path ${configFile.canonicalPath}")
-    }
-    try {
-        return SpecmaticConfigMapper().read(configFile)
-    } catch(e: LinkageError) {
-        logger.log(e, "A dependency version conflict has been detected. If you are using Spring in a maven project, a common resolution is to set the property <kotlin.version></kotlin.version> to your pom project.")
-        throw e
-    } catch (e: JsonParseException) {
-        val msg = "Invalid configuration file format: " + e.message
-        logger.log(e, msg)
-        throw IllegalArgumentException(msg, e)
-    } catch (e: JsonMappingException) {
-        val msg = "Your configuration file does not match the expected structure: " + e.message
-        logger.log(e, msg)
-        throw IllegalArgumentException(msg, e)
-    } catch (e: IOException) {
-        val msg = "Error reading configuration file: " + e.message
-        logger.log(e, msg)
-        throw RuntimeException(msg, e)
-    } catch (e: Throwable) {
-        logger.log(e, "Your configuration file may have some missing configuration sections. Please ensure that the ${configFile.path} file adheres to the schema described at: https://specmatic.io/documentation/specmatic_json.html")
-        throw Exception("Your configuration file may have some missing configuration sections. Please ensure that the ${configFile.path} file adheres to the schema described at: https://specmatic.io/documentation/specmatic_json.html", e)
     }
 }
