+++ core/src/main/kotlin/io/specmatic/core/config/SpecmaticConfigConstants.kt
@@ -1,0 +1,22 @@
+package io.specmatic.core.config
+
+const val APPLICATION_NAME = "Specmatic"
+const val APPLICATION_NAME_LOWER_CASE = "specmatic"
+const val CONFIG_FILE_NAME_WITHOUT_EXT = "specmatic"
+const val DEFAULT_TIMEOUT_IN_MILLISECONDS: Long = 6000L
+const val CONTRACT_EXTENSION = "spec"
+const val YAML = "yaml"
+const val WSDL = "wsdl"
+const val YML = "yml"
+const val JSON = "json"
+val CONFIG_EXTENSIONS = listOf(YAML, YML, JSON)
+val OPENAPI_FILE_EXTENSIONS = listOf(YAML, YML, JSON)
+val CONTRACT_EXTENSIONS = listOf(CONTRACT_EXTENSION, WSDL) + OPENAPI_FILE_EXTENSIONS
+const val DATA_DIR_SUFFIX = "_data"
+const val TEST_DIR_SUFFIX = "_tests"
+const val EXAMPLES_DIR_SUFFIX = "_examples"
+const val DICTIONARY_FILE_SUFFIX = "_dictionary.json"
+const val SPECMATIC_GITHUB_ISSUES = "https://github.com/znsio/specmatic/issues"
+const val DEFAULT_WORKING_DIRECTORY = ".$APPLICATION_NAME_LOWER_CASE"
+const val SPECMATIC_STUB_DICTIONARY = "SPECMATIC_STUB_DICTIONARY"
+const val MISSING_CONFIG_FILE_MESSAGE = "Config file does not exist. (Could not find file ./specmatic.json OR ./specmatic.yaml OR ./specmatic.yml)"
